refactor: userId를 Long 자료형으로 변경

+++ src/main/java/forif/univ_hanyang/apply/dto/request/IsPaidRequest.java
@@ -6,6 +6,6 @@ import java.util.Set;
 
 @Data
 public class IsPaidRequest {
-    private Set<Integer> applierIds;
+    private Set<Long> applierIds;
     private String payYn;
 }
